requirements.txt
selenium>=4.10.0
pandas>=2.0.0
streamlit>=1.24.0
webdriver-manager>=4.0.0
pyyaml>=6.0.0
